cv_parser/requirements.txt
pdfminer.six
python-docx
pytest